Backend/requirements.txt
aiofiles==24.1.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
asn1crypto==1.5.1
async-property==0.2.2
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bz2file==0.98
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
concurrent-log-handler==0.9.25
contourpy==1.3.2
cryptography==44.0.1
cycler==0.12.1
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
entrypoints==0.4
et_xmlfile==2.0.0
extra-streamlit-components==0.1.80
Faker==37.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi_keycloak==1.1.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kiwisolver==1.4.8
linkify-it-py==2.0.3
lxml==5.4.0
lz4==4.4.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.7.0
mypy==1.1.1
mypy_extensions==1.1.0
narwhals==1.36.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.1.4
passlib==1.7.4
patsy==1.0.1
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.7
plotly==6.0.1
portalocker==3.1.1
prettytable==3.7.0
progress==1.6
prometheus_client==0.21.1
protobuf==5.29.4
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.22
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.3
pyodbc==5.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-keycloak==5.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.1
referencing==0.36.2
reportlab==4.2.2
requests==2.32.3
requests-extensions==1.1.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.3
rpds-py==0.24.0
rsa==4.9.1
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==1.4.46
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.46.2
statsmodels==0.14.4
stauth==0.1.12
streamlit==1.44.1
streamlit-aggrid==1.1.4.post1
streamlit-authenticator==0.4.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-javascript==0.1.5
streamlit-keyup==0.3.0
streamlit-pydantic==0.6.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
thrift==0.16.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
trino==0.327.0
typer==0.15.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.34.2
validators==0.33.0
watchdog==4.0.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
XlsxWriter==3.2.0
